+++ src/DOMSelector/Exceptions/InvalidTypeException.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DOMSelector\Exceptions;
+
+use Exception;
+
+class InvalidTypeException extends Exception
+{
+    //
+}