+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [Unreleased](https://github.com/ytekeli/dom-selector/compare/v0.1.0...master)
 
 ### Added
+- Add `TypeProvider` feature for including your own types to selector from out of instance.
 - Add `extractFromFile` and `extractFromUrl` features.
 - Improve some test cases.